+++ src/sender/game/index.ts
@@ -1,0 +1,6 @@
+export * from './AddMatrix';
+export * from './Attack';
+export * from './RandomAttack';
+export * from './StartGame';
+export * from './Killed';
+export * from './WhoMove';
