+++ src/sender/updater/index.ts
@@ -1,0 +1,2 @@
+export * from './UpdateRooms';
+export * from './UpdateWinners';
