+++ src/utils/FindGame.ts
@@ -1,0 +1,12 @@
+import { indexes } from '../data';
+
+export function getGameIdByPlayerIndex(indexPlayer: string): number | undefined {
+    const data = indexes.find((user) => user.idPlayer === indexPlayer);
+    
+    if (!data) {
+      return undefined;
+    }
+    
+    return data.idGame;
+  }
+
