+++ src/utils/SendToAll.ts
@@ -1,0 +1,6 @@
+import { CustomWebSocket, Request } from '../types/index';
+export const sendToAllClients = (message: Request, wsclients:CustomWebSocket[]) =>{
+    wsclients.forEach((client) => {
+        client.send(JSON.stringify(message));
+    });
+};
