+++ src/utils/WhiteSpaceCheck.ts
@@ -1,0 +1,3 @@
+export function hasWhitespace(str: string): boolean {
+    return /\s/.test(str);
+}
